% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.database.connect.R
\name{check.conn}
\alias{check.conn}
\title{Check and Validate Database Connection}
\usage{
check.conn(conn = NULL, db.credentials = NULL)
}
\arguments{
\item{conn}{A DBI connection object. If \code{NULL}, the function attempts to find a valid connection
in the global environment or initialize a new connection.}

\item{db.credentials}{A list of database credentials including user, password, host, and port.
If \code{NULL}, credentials are fetched using \code{get.credentials}.}
}
\value{
A valid DBI connection object.
}
\description{
This function verifies the validity of a given database connection. If the connection is not valid
or not provided, it attempts to retrieve an existing connection from the global environment. If no
valid connection is found, it initializes a new connection using the provided database credentials.
}
\examples{
\dontrun{
  # Check an existing connection or create a new one
  credentials <- list(
    BIAD_DB_USER = "yourusername",
    BIAD_DB_PASS = "yourpassword",
    BIAD_DB_HOST = "127.0.0.1",
    BIAD_DB_PORT = 3306
  )
  
  conn <- check.conn(conn = NULL, db.credentials = credentials)
}

}
