% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create.html.for.row.comments}
\alias{create.html.for.row.comments}
\title{Create an HTML file : should be move to BIADWIKI}
\usage{
create.html.for.row.comments(table.data, column.data, file)
}
\description{
Create an HTML file : should be move to BIADWIKI
}
