% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create.html.for.table.comments}
\alias{create.html.for.table.comments}
\title{Create an HTML Table with Comments should be move to BIADWIKI}
\usage{
create.html.for.table.comments(table.data, column.data, file)
}
\arguments{
\item{table.data}{A data frame containing information about tables including their names and row counts.}

\item{column.data}{A data frame containing information about table columns.}

\item{file}{A string specifying the path to the output HTML file.}
}
\value{
This function does not return a value. It creates and saves an HTML file as a side effect.
}
\description{
This function generates an HTML table summarizing the number of rows, columns, and column names for a given set of tables.
It utilizes the \code{gt} package to format the table and saves it to a specified HTML file.
}
