% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{database.relationship.plotter}
\alias{database.relationship.plotter}
\title{Plot Database Table Relationships}
\usage{
database.relationship.plotter(
  d.tables,
  include.look.ups = TRUE,
  conn = NULL,
  db.credentials = NULL
)
}
\arguments{
\item{d.tables}{A character vector of table names to include in the plot, separated by semicolons and spaces (e.g., "table1; table2").}

\item{include.look.ups}{Logical. If TRUE, lookup tables and their relationships are included in the plot. Default is TRUE.}

\item{conn}{A database connection object. If NULL, \code{db.credentials} is used to establish a connection.}

\item{db.credentials}{A list containing database credentials, such as username, password, and database name. Used to establish a connection if \code{conn} is NULL.}
}
\value{
A DiagrammeR graph object representing the database schema as a visual diagram.
}
\description{
helper to visualizes the relationships between all the tables (used for the wiki)
}
\details{
The function queries the \code{INFORMATION_SCHEMA.KEY_COLUMN_USAGE} to obtain foreign key relationships between tables in the database schema named 'BIAD'. It then uses DiagrammeR to create a visual representation of these relationships.
}
