% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.ancestors}
\alias{get.ancestors}
\title{Retrieve Ancestor Records from Database}
\usage{
get.ancestors(
  keys,
  table.name,
  primary.value,
  conn = NULL,
  db.credentials = NULL,
  orig.table = NULL,
  zoption = FALSE
)
}
\arguments{
\item{keys}{A data frame containing database information, including relationships between tables (obtained via \code{get.keys})}

\item{table.name}{A string specifying the name of the table from which to start retrieving descendant records.}

\item{primary.value}{The primary key value from which to find descendant records.}

\item{conn}{A database connection object.}

\item{db.credentials}{manual database credentials.}

\item{orig.table}{a parameter that store the orinal table at the level of the row matching the object. Used for the recursion}

\item{zoption}{if TRUE, zoptions tables linked to the object  are also returned. Default is \code{FALSE}.!!NOT IMPLEMENTED}
}
\value{
A nested list containing data frames of descendant records for each related table.
}
\description{
This function retrieves all ancestor records related to a specified primary value in a database table.
}
