% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.database.connect.R
\name{get.credentials}
\alias{get.credentials}
\title{Retrieve Credentials from Environment Variables}
\usage{
get.credentials()
}
\value{
A list containing the database user, password, host, and port.
}
\description{
This function fetches database credentials from specified environment variables.
}
