% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.primary.column.from.table}
\alias{get.primary.column.from.table}
\title{Get Primary Key Column from Table}
\usage{
get.primary.column.from.table(
  keys = NULL,
  table.name,
  conn = NULL,
  db.credentials = NULL
)
}
\arguments{
\item{keys}{A data frame containing key information for database tables. If \code{NULL}, keys are retrieved from the database using the \code{get.keys} function.}

\item{table.name}{A character string specifying the name of the table for which the primary key column is being queried.}

\item{conn}{A DBI connection object. If \code{NULL}, the function will attempt to check or initialize a  connection using the provided or default \code{db.credentials}.}

\item{db.credentials}{A list containing database connection details, including user, password, host, and port. If \code{NULL}, credentials are fetched using \code{get.credentials}.}
}
\value{
A character string representing the name of the primary key column. If no primary key is found, \code{NA} is returned. If multiple potential primary keys are found, the function stops with an error.
}
\description{
This function retrieves the name of the primary key column from a specified table in the database.
It first checks for any unique or primary constraints, and then determines which column serves as
the primary key. If there is ambiguity in determining the primary key, the function will stop with an error.
}
