% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.relatives}
\alias{get.relatives}
\title{Retrieve Relatives from Database Table}
\usage{
get.relatives(
  table.name,
  primary.value,
  directions = c("up", "down"),
  conn = NULL,
  db.credentials = NULL,
  zoption = FALSE
)
}
\arguments{
\item{table.name}{A string specifying the name of the table where the entry is.}

\item{primary.value}{The primary key value used to find the entry in the database.}

\item{directions}{A character vector indicating direction(s) for retrieving the data related to the entry
Available options are "up" for ancestors and "down" for descendants. Default is both (\code{directions=c("up", "down")}).}

\item{conn}{A database connection object. Default is \code{NULL}.}

\item{db.credentials}{parameter for manual setup of database credentials. Default is \code{NULL}.}

\item{zoption}{if TRUE, zoptions tables linked to the object  are also returned. Default is \code{FALSE}.}
}
\value{
A list containing a root element with one branch with all the data associated with the specific entry and two other branches storing trees as nested list with all related entries.
}
\description{
This function generates trees of ancestor or descendant records related to a specific entries a database table.
}
