% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.table.data}
\alias{get.table.data}
\title{Retrieve Table Entries from Database}
\usage{
get.table.data(
  keys = NULL,
  table.name = NULL,
  primary.value = NULL,
  conn = NULL,
  db.credentials = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{keys}{A vector of key names used to determine the primary column of the table.}

\item{table.name}{A string specifying the name of the table from which to retrieve data.}

\item{primary.value}{A value or a vector of values that are used to filter the rows in the table based on the primary column.}

\item{conn}{A database connection object to be used for the query. If NULL, db.credentials should be provided.}

\item{db.credentials}{Credentials required to establish a database connection, used when conn is NULL.}

\item{na.rm}{A logical value indicating whether to remove columns with all NA values from the result. The default is TRUE.}
}
\value{
A data frame containing the queried data, potentially with NA columns removed.
}
\description{
This function queries a database table to retrieve information about one or multiple entries in the database
}
