% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_elements}
\alias{get_elements}
\title{Get Elements from a Tree Structure}
\usage{
get_elements(x, element)
}
\arguments{
\item{x}{A list representing the tree structure.}

\item{element}{A character string specifying the name of elements to extract from the tree.}
}
\value{
A list containing all elements from the tree that have names matching the specified element.
}
\description{
This function extracts from a tree, as created by the function \code{get.relatives},
all nodes/leaves that have names corresponding to the specified element.
trees should be as named list of list like list(a=1,b=list(a=2,b=3,c=4),d=list(e=1,f=4))
Based on the answers here: https://stackoverflow.com/questions/64578972/pull-all-elements-with-specific-name-from-a-nested-list/79168230#79168230
}
\examples{
\dontrun{
tree <- list(a = 1, b = list(a = 2, c = 3), d = 4)
get_elements(tree, "a")
# Expected output: list(1, 2)
}
}
