% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.database.connect.R
\name{init.conn}
\alias{init.conn}
\title{Initialize Database Connection}
\usage{
init.conn(db.credentials = NULL, dbname = "BIAD")
}
\arguments{
\item{db.credentials}{A list containing database connection details. The list
should include \code{user}, \code{password}, \code{host}, and \code{port}. If \code{NULL}, defaults
to fetching these values from environment variables. You can store these in
\verb{~/.Renviron} or export them in your environment using your favorite method
(ie: $export host='127.0.0.1')}
}
\value{
A DBI connection object to the MySQL database.
}
\description{
This function initializes a connection to the BIAD database using the provided
credentials. If no credentials are supplied, it attempts to retrieve them from
environment variables that should be stored in \verb{~/.Renviron}
}
