% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make.database.user}
\alias{make.database.user}
\title{Set up new BIAD user NOT SURE WHERE TO PUT THIS FUNCTION YET}
\usage{
make.database.user(databases, username, email, team, committee, administrator)
}
\arguments{
\item{databases}{A character vector of length >=1, giving names of the databases to give access permission to. Typically 'BIAD'.}

\item{username}{A character vector of length = 1. Ideally all lowercase, first name and surname joined with an underscore.}

\item{email}{A character vector of length =1, giving a valid email address.}

\item{team}{A Either 'YES' or 'NO'.}

\item{committee}{A Either 'YES' or 'NO'.}

\item{administrator}{A Either 'YES' or 'NO'.}
}
\value{
This function does not return a value. It performs various changes to the database, and creates and saves 'email.txt'.
}
\description{
This function is a headstart setting up a new user, or changing details for an existing user.
It stores username, email and credential level in the database, writes a draft email (sending will be automated eventually),
and creates the user credentials in the database, including a random generated password.
}
