% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.database.connect.R
\name{msp}
\alias{msp}
\title{msp(Mask Password)
This function masks a given password by replacing all but the first and last character with asterisks.}
\usage{
msp(password)
}
\arguments{
\item{password}{A character string representing the password to be masked.}
}
\value{
A character string with the masked password.
}
\description{
msp(Mask Password)
This function masks a given password by replacing all but the first and last character with asterisks.
}
