% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{slc}
\alias{slc}
\title{Shortest Distance Calculation Using Spherical Law of Cosines}
\usage{
slc(x, y, ax, ay, input = "rad")
}
\arguments{
\item{x}{A numeric value representing the longitude of the single point in radians or degrees.}

\item{y}{A numeric value representing the latitude of the single point in radians or degrees.}

\item{ax}{A numeric vector representing the longitudes of the point set in radians or degrees.}

\item{ay}{A numeric vector representing the latitudes of the point set in radians or degrees.}

\item{input}{A character string indicating the input unit. It can be either \code{'rad'} for radians
(default) or \code{'deg'} for degrees.}
}
\value{
A numeric vector of distances in radians from the single point \code{(x, y)} to each of the points in \code{(ax, ay)}.
}
\description{
This function calculates the shortest distance between a single point \code{(x, y)} and a set of points \code{(ax, ay)}
using the spherical law of cosines. The distances are returned in radians and can be converted to kilometers by
multiplying by the Earth's radius (6378.1 km).
}
\details{
The function requires input coordinates in either radians or degrees. If the input is in degrees, they will be
converted to radians internally. The function ensures that the cosine law calculation stays within its valid range to avoid floating point inaccuracies.
}
