% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.maker}
\alias{summary.maker}
\title{Summary Maker for Site Data}
\usage{
\method{summary}{maker}(d)
}
\arguments{
\item{d}{dataframe of site, with SiteID}
}
\description{
Summary Maker for Site Data
}
