% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary_maker}
\alias{summary_maker}
\title{Summary Maker for Site Data}
\usage{
summary_maker(d)
}
\arguments{
\item{d}{dataframe of site, with SiteID}
}
\value{
a smumarry
}
\description{
Summary Maker for Site Data
}
